requirements.txt
torch>=1.9.0
transformers>=4.20.0
pandas>=1.3.0
numpy>=1.21.0
scikit-learn>=1.0.0
elasticsearch>=8.0.0
python-dotenv>=0.19.0
psycopg2-binary>=2.9.0
pgvector>=0.1.0
requests>=2.28.0

